requirements_gui.txt
# SPMS Requirements - Updated with Web GUI

# Core dependencies
sqlalchemy>=2.0.0
pyyaml>=6.0

# Web GUI dependencies
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
